Move publications to subdirectory
My publications need to be better organized.  For this, a subdirectory
with its own SConscript is a good idea.  Now I have a filter to break
the types into subsections and I can get the full list to add to my CV.
As a note, I prefer using the DOI instead of the direct link from the
publisher if it is available.