Update the CSS colors and structure
Now we use a color scheme loosely based on Zenburn.  We also update the
main style sheet to use variables for the colors in case we ever want to
change the color scheme.

I also adjusted some of the other structural elements to tighten up the
look including using ems to set the size.