Add On Campus Committees
This is the section after the professional activities.  Luckily, it has
the same format as the professional activities meaning we can just reuse
the previous script.  We do need to update it to accept a key with which
to search.  Other than that, we just need to add new data to the YAML
file.