Add utilities for working with my projects
We want to import my project listing from a common file.  To help with
this, I want to start migrating to using a package for managing the
scripts.  The underlying problem is I'm tired of repeating some
functions like `find`.  These two scripts allow me to run using

    python -m resume.projects ...
    python -m resume.projects.import ...

Life should be better now.