Comment out lwarp details
I'm not sure this will be the way to go.  It does not look like lwarp
correctly deals with `lstinputfile`, and I'm not sure I want to deal
with all the extra files in the tree it generates.  This is where I left
off in case I decide to actually go back down this path.