Fix the template for the website
I wanted to add the abstract to the final output.  This means I had to
update Pandoc to a newer version mainly because the version of Pandoc I
had generated the title, author, and date as part of the body.  The
newer template gave me a place to put the abstract.  But, for some
reason, Pandoc was not respecting the --data-dir flag (even after I
corrected it).  So, instead we just use an explicit template.