Fix the computational physics post
It needed a log of updating to get the math rendering correctly and the
images in place.  I did discover that I need to get the Pandoc Tool to
scan the resource path.  At the moment, it only looks for the included
images in the _output_ directory.  If given the `resource-path` (like
this post needs), Pandoc will search on the path to locate resources.