Purge old Jekyll based build files
In the process of migrating to Sphinx and my theme, I really just want
to get rid of the files that are no longer relevant.  This includes the
files for building using Jekyll and the accumulated cruft from the
transition to a separate repository (which I just merged back over
here).