Reorganize the fun stuff to using reStructuredText
This just moves the reorganizes the files to the proper location.  I
might come back and actually work on these some more, but probably not.
The spell list generation could be a fun experiment in setting up an
extension that generates a source, but it's not like I'll be playing 3.5
any time soon.  And the character sheet is pretty much a lost cause at
this point.